# docs/swagger.yaml
basePath: /
definitions:
...
  github_com_patrick_cocobase_internal_models.AggregateResponse:
    properties:
      collection:
        type: string
      field:
        type: string
      operation:
        type: string
      result: {}
    type: object
...
  github_com_patrick_cocobase_internal_models.CountResponse:
    properties:
      count:
        type: integer
    type: object
...
  github_com_patrick_cocobase_internal_models.FieldInfo:
    properties:
      null_percentage:
        type: number
      nullable:
        type: boolean
      primary_type:
        type: string
      samples:
        items: {}
        type: array
      types:
        items:
          type: string
        type: array
    type: object
...
  github_com_patrick_cocobase_internal_models.RelationshipInfo:
    properties:
      confidence:
        type: string
      field:
        type: string
      target_collection:
        type: string
      type:
        type: string
    type: object
  github_com_patrick_cocobase_internal_models.SchemaResponse:
    properties:
      analyzed_documents:
        type: integer
      collection:
        type: string
      detected_relationships:
        additionalProperties:
          $ref: '#/definitions/github_com_patrick_cocobase_internal_models.RelationshipInfo'
        type: object
      document_count:
        type: integer
      fields:
        additionalProperties:
          $ref: '#/definitions/github_com_patrick_cocobase_internal_models.FieldInfo'
        type: object
      usage_hint:
        type: string
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@cocobase.io
    name: Cocobase Support
  description: Backend as a Service with flexible collections and document management
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Cocobase API
  version: "1.0"
paths:
...
  /collections/{id}/query/documents/aggregate:
    get:
      description: Perform aggregation operations (count, sum, avg, min, max) on document
        fields
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      - description: Field to aggregate
        in: query
        name: field
        required: true
        type: string
      - default: count
        description: Aggregation operation (count, sum, avg, min, max)
        in: query
        name: operation
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_patrick_cocobase_internal_models.AggregateResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Aggregate documents
      tags:
      - Advanced Queries
  /collections/{id}/query/documents/count:
    get:
      description: Count documents matching filters without returning the documents.
        Useful for pagination and statistics.
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_patrick_cocobase_internal_models.CountResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Count documents
      tags:
      - Advanced Queries
  /collections/{id}/query/documents/group-by:
    get:
      description: Group documents by a field and count occurrences
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      - description: Field to group by
        in: query
        name: field
        required: true
        type: string
      - description: 'Field to count (default: document count)'
        in: query
        name: count_field
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Group by field
      tags:
      - Advanced Queries
  /collections/{id}/query/schema:
    get:
      description: Analyze collection and return inferred schema with field types
        and detected relationships
      parameters:
      - description: Collection ID or Name
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_patrick_cocobase_internal_models.SchemaResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: Get collection schema
      tags:
      - Advanced Queries
...
schemes:
- http
- https
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-API-Key
    type: apiKey
  BearerAuth:
    description: Type "Bearer" followed by a space and API key.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"